lehmer extended gcd

+++ ecc/src/integer_arithmetic/addition.rs
@@ -16,11 +16,11 @@ impl Addition for BigInteger {
             basis: *b,
         };
         let max_n = std::cmp::max(nu, nv);
-        let mut carrier = 0 as i32;
+        let mut carrier = 0 as u8;
         // from right to left
         for i in 0..max_n {
-            let lft = if i <= nu - 1 { u[i] } else { 0 as i32 };
-            let rht = if i <= nv - 1 { v[i] } else { 0 as i32 };
+            let lft = if i <= nu - 1 { u[i] } else { 0 as u8 };
+            let rht = if i <= nv - 1 { v[i] } else { 0 as u8 };
             let t = lft + rht + carrier;
             let remainder = if t >= *b { t - b } else { t };
             carrier = if t >= *b { 1 } else { 0 };
